refactor: 3단계 로또(2등) - lottoGame 패키지명을 lottogame로 개선 - 당첨번호,보너스볼을 관리하는 WinningLottoNums 객체 생성 - 당첨결과 추출하는 행위를 Lotto에서 WinningLottoNums로 이관
