--- src/main/java/lottoGame/Application.java
@@ -1,10 +1,0 @@
-package lottoGame;
-
-import lottoGame.controller.LottoStore;
-
-public class Application {
-
-    public static void main(String[] args) {
-        new LottoStore().start();
-    }
-}
