+++ src/main/java/lottogame/Application.java
@@ -1,0 +1,10 @@
+package lottogame;
+
+import lottogame.controller.LottoStore;
+
+public class Application {
+
+    public static void main(String[] args) {
+        new LottoStore().start();
+    }
+}
