+++ src/main/java/lottogame/model/lotto/Lottos.java
@@ -1,0 +1,37 @@
+package lottogame.model.lotto;
+
+import java.util.List;
+import lottogame.model.winner.WinStandard;
+import lottogame.model.winner.WinnerResult;
+
+public class Lottos {
+
+    private final List<Lotto> lottos;
+
+    public Lottos(List<Lotto> lottos) {
+        this.lottos = lottos;
+    }
+
+    public WinnerResult compareAndElectWinResult(
+            final WinningLottoNums winningLottoNums
+    ) {
+        WinnerResult winnerResult = new WinnerResult();
+        for (Lotto lotto : lottos) {
+            WinStandard winStandard = winningLottoNums.checkIfWin(lotto);
+//            WinStandard winStandard = lotto.checkIfWin(beforeWinLotto, bonusNum);
+            winnerResult.addWinResult(winStandard);
+        }
+
+        return winnerResult;
+    }
+
+    public int size() {
+        return this.lottos.size();
+    }
+
+    public List<String> convertRawString() {
+        return this.lottos.stream()
+                .map(Lotto::toString)
+                .toList();
+    }
+}
